# docs/swagger.yaml
openapi: 3.0.2
info:
    title: games-api
    version: 1.0.0
    description: This is a Swagger UI for our Games API
paths:
    /games:
        get:
            tags:
                - games
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: string
                            examples:
                                List of games:
                                    value:
                                        - Witcher 3
                                        - Cyberpunk 2077
                    description: application/json
            summary: Get list of all games
            description: Returns a list of all games in API
        post:
            requestBody:
                description: A new game to be created
                required: true
            tags:
                - new-game
            responses:
                '201':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/new-game'
                    description: create a new game
                '400':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                    description: One or all params are missing
            summary: add game in list
            description: creates a new game
    '/games/{id}':
        get:
            tags:
                - details
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                Full details of a given game:
                                    value: |-
                                        {
                                            "id": 1,
                                            "name": "Witcher 2",
                                            "price": 29.99
                                          },
                                          {
                                            "id": 2,
                                            "name": "Cyberpunk 2077",
                                            "price": 59.99
                                          },
                                          {
                                            "id": 3,
                                            "name": "Minecraft",
                                            "price": 26.99
                                          }
                    description: Returns full details of a given id game
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                Game \\w+ was not found in game array:
                                    value: Game was not found in game array.
                    description: Game was not found in games array
            operationId: id
            summary: Get full details of a game
            description: Returns full details of a given id game
        parameters:
            -
                name: id
                in: path
                required: true
components:
    schemas:
        error:
            title: Root Type for error
            description: ''
            type: object
            properties:
                error:
                    type: string
            example:
                error: Error message.
        new-game:
            title: Root Type for new-game
            description: ''
            required:
                - price
                - name
            type: object
            properties:
                name:
                    type: string
                price:
                    format: float
                    type: number
            example:
                name: Tetris
                price: '0.99'